d3d: Begin centralizing framebuffer management.
