+++ GPU/Directx9/FramebufferDX9.h
@@ -30,57 +30,12 @@
 
 #include "Globals.h"
 #include "GPU/GPUCommon.h"
+#include "GPU/Common/FramebufferCommon.h"
 
 namespace DX9 {
 
 struct GLSLProgram;
 class TextureCacheDX9;
-
-enum {
-	FB_USAGE_DISPLAYED_FRAMEBUFFER = 1,
-	FB_USAGE_RENDERTARGET = 2,
-	FB_USAGE_TEXTURE = 4,
-};
-
-enum {	
-	FB_NON_BUFFERED_MODE = 0,
-	FB_BUFFERED_MODE = 1,
-	FB_READFBOMEMORY_CPU = 2,
-	FB_READFBOMEMORY_GPU = 3,
-};
-
-struct VirtualFramebufferDX9 {
-	int last_frame_used;
-	int last_frame_attached;
-	int last_frame_render;
-	bool memoryUpdated; 
-
-	u32 fb_address;
-	u32 z_address;
-	int fb_stride;
-	int z_stride;
-
-	// There's also a top left of the drawing region, but meh...
-
-	// width/height: The detected size of the current framebuffer.
-	u16 width;
-	u16 height;
-	// renderWidth/renderHeight: The actual size we render at. May be scaled to render at higher resolutions.
-	u16 renderWidth;
-	u16 renderHeight;
-	// bufferWidth/bufferHeight: The actual (but non scaled) size of the buffer we render to. May only be bigger than width/height.
-	u16 bufferWidth;
-	u16 bufferHeight;
-
-	u16 usageFlags;
-
-	GEBufferFormat format;  // virtual, right now they are all RGBA8888
-	FBOColorDepth colorDepth;
-	FBO *fbo;
-
-	bool dirtyAfterDisplay;
-	bool reallyDirtyAfterDisplay;  // takes frame skipping into account
-};
 
 void CenterRect(float *x, float *y, float *w, float *h,
 								float origW, float origH, float frameW, float frameH);
@@ -88,7 +43,7 @@ void CenterRect(float *x, float *y, float *w, float *h,
 
 class ShaderManagerDX9;
 
-class FramebufferManagerDX9 {
+class FramebufferManagerDX9 : public FramebufferManagerCommon {
 public:
 	FramebufferManagerDX9();
 	~FramebufferManagerDX9();
@@ -112,62 +67,41 @@ public:
 	void Resized();
 	void DeviceLost();
 	void CopyDisplayToOutput();
-	void SetRenderFrameBuffer();  // Uses parameters computed from gstate	
+	virtual void DoSetRenderFrameBuffer() override;  // Uses parameters computed from gstate
 	void UpdateFromMemory(u32 addr, int size, bool safe);
 
-	void ReadFramebufferToMemory(VirtualFramebufferDX9 *vfb, bool sync = true);
+	void ReadFramebufferToMemory(VirtualFramebuffer *vfb, bool sync = true);
 
 	// TODO: Break out into some form of FBO manager
-	VirtualFramebufferDX9 *GetVFBAt(u32 addr);
-	VirtualFramebufferDX9 *GetDisplayVFB() {
+	VirtualFramebuffer *GetVFBAt(u32 addr);
+	VirtualFramebuffer *GetDisplayVFB() {
 		return GetVFBAt(displayFramebufPtr_);
 	}
-	void SetDisplayFramebuffer(u32 framebuf, u32 stride, GEBufferFormat format);
-	size_t NumVFBs() const { return vfbs_.size(); }
 
 	std::vector<FramebufferInfo> GetFramebufferList();
 
-	int GetRenderWidth() const { return currentRenderVfb_ ? currentRenderVfb_->renderWidth : 480; }
-	int GetRenderHeight() const { return currentRenderVfb_ ? currentRenderVfb_->renderHeight : 272; }
-	int GetTargetWidth() const { return currentRenderVfb_ ? currentRenderVfb_->width : 480; }
-	int GetTargetHeight() const { return currentRenderVfb_ ? currentRenderVfb_->height : 272; }
-
-	u32 PrevDisplayFramebufAddr() {
-		return prevDisplayFramebuf_ ? (0x04000000 | prevDisplayFramebuf_->fb_address) : 0;
-	}
-	u32 DisplayFramebufAddr() {
-		return displayFramebuf_ ? (0x04000000 | displayFramebuf_->fb_address) : 0;
-	}
-
 	void NotifyFramebufferCopy(u32 src, u32 dest, int size);
 
-	void DestroyFramebuf(VirtualFramebufferDX9 *vfb);
+	void DestroyFramebuf(VirtualFramebuffer *vfb);
 
 	bool GetCurrentFramebuffer(GPUDebugBuffer &buffer);
 	bool GetCurrentDepthbuffer(GPUDebugBuffer &buffer);
 	bool GetCurrentStencilbuffer(GPUDebugBuffer &buffer);
+
+protected:
+	virtual void DisableState() override;
+	virtual void ClearBuffer() override;
+	virtual void ClearDepthBuffer() override;
 
 private:
 	void CompileDraw2DProgram();
 	void DestroyDraw2DProgram();
 
 	void SetNumExtraFBOs(int num);
-	u32 displayFramebufPtr_;
-	u32 displayStride_;
-	GEBufferFormat displayFormat_;
-
-	VirtualFramebufferDX9 *displayFramebuf_;
-	VirtualFramebufferDX9 *prevDisplayFramebuf_;
-	VirtualFramebufferDX9 *prevPrevDisplayFramebuf_;
-	int frameLastFramebufUsed;
-
-	std::vector<VirtualFramebufferDX9 *> vfbs_;
-
-	VirtualFramebufferDX9 *currentRenderVfb_;
 
 	// Used by ReadFramebufferToMemory
-	void BlitFramebuffer_(VirtualFramebufferDX9 *src, VirtualFramebufferDX9 *dst, bool flip = false, float upscale = 1.0f, float vscale = 1.0f);
-	void PackFramebufferDirectx9_(VirtualFramebufferDX9 *vfb);
+	void BlitFramebuffer_(VirtualFramebuffer *src, VirtualFramebuffer *dst, bool flip = false, float upscale = 1.0f, float vscale = 1.0f);
+	void PackFramebufferDirectx9_(VirtualFramebuffer *vfb);
 	
 	// Used by DrawPixels
 	LPDIRECT3DTEXTURE9 drawPixelsTex_;
@@ -188,7 +122,7 @@ private:
 	bool resized_;
 	bool useBufferedRendering_;
 
-	std::vector<VirtualFramebufferDX9 *> bvfbs_; // blitting FBOs
+	std::vector<VirtualFramebuffer *> bvfbs_; // blitting FBOs
 
 	std::set<std::pair<u32, u32>> knownFramebufferCopies_;
