+++ GPU/GPU.vcxproj.filters
@@ -171,6 +171,9 @@
     <ClInclude Include="GLES\FragmentTestCache.h">
       <Filter>GLES</Filter>
     </ClInclude>
+    <ClInclude Include="Common\FramebufferCommon.h">
+      <Filter>Common</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
@@ -322,6 +325,9 @@
     </ClCompile>
     <ClCompile Include="GLES\FragmentTestCache.cpp">
       <Filter>GLES</Filter>
+    </ClCompile>
+    <ClCompile Include="Common\FramebufferCommon.cpp">
+      <Filter>Common</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
